feat: moved permutator to dedicated file. Elided some lifetimes
